Move solo strategy into its own file
